backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
boto3
qdrant-client
python-multipart
fastembed
onnxruntime
Pillow
redis
